+++ src/main/java/mil/nga/geopackage/extension/nga/link/FeatureTileLinkDao.java
@@ -184,7 +184,7 @@ public class FeatureTileLinkDao
 
 	/**
 	 * {@inheritDoc}
-	 * 
+	 * <p>
 	 * Update using the complex key
 	 */
 	@Override